Remove _attr_device_class
Due to a misunderstanding as to exactly what this attribute actually does and changes in 2023.x this attribute no longer changes the original id of the entity and only seeks to tweak the entity id in weird ways.  The _attr_has_entity_name is the only thing that can be used to tweak the id for the entity.